MVP, interactive filtering works, show #cells left
A bit of a dirty commit. Went back a bit, started doing several things
at once: figuring out reactive connections, breaking out code into
modules, etc. So moved back and worked on "monolith-code" for a while,
to get the reactive stuff working for a single metadata column, then it
should be easier to break out and generalise for `n` number of columns.