added back dir load button, works and updates UI
rudimentary checks are performed for correctness of the selected
directory. Namely, if the dir exists and contains the files
in first the set: `{matrix.mtx,features.tsv,barcodes.tsv}.gz`
and if not those the files in the set
`{matrix.mtx,features.tsv,barcodes.tsv}`. If not, UI informing the user
that the selected dir is not a valid count matrix directory is shown.
If the checks above are correct, UI informing of correctness is shown,
and a (currently placeholder) message that the data is being loaded into
`Seurat`.

Could have been in an earlier commit as well, but the `reactiveValues`
list (which is my current best implementation of a single source of
truth) is shown in the sidebar, mostly for debugging purposes atm.