fairly big change: possible to run from server
The gist of it: the shinyFiles package was only for a server-side file
system. Which is entirely fine if the Shiny app is run from the same
computer the intended data files are on, but any other case (e.g.
serving the app from somewhere), this doesn't make sense.

A _load_ of dependencies, very annoying to update smth like this. Seems
hard to decouple e.g. UI, backend logic etc. with Shiny. I tried in the
beginning with modules, but that seemed to change the reactivity
response times too much. So this commit felt like poking a big pile of
sticky spaghetti. Fun. But hey, now it works. Still unbearably
rudimentary though, basically checks for if the uploaded file names
match the base case of count matrix files (i.e.
`{matrix.mtx,{features,genes}.tsv,barcodes.tsv}{,.gz}`). And not much
more. But, hey, it works.. In that specific case.